camera coordinate works now. All coordinates are kept inside cameraCoordinate.js
