+++ vis/scripts/cameraCoordinate.js
@@ -1,0 +1,17 @@
+export let animals =[
+    {"name": "Lion",
+     "cameraX": 0,
+     "cameraY": 5,
+     "cameraZ": 0
+    },
+    {"name": "Camera 2",
+     "cameraX": -90,
+     "cameraY": 0,
+     "cameraZ": 0,
+    },
+    {"name": "Camera 3",
+     "cameraX": 80,
+     "cameraY": 0,
+     "cameraZ": 0,
+    },
+]
