+++ vis/scripts/cameraNavigation.js
@@ -1,46 +1,24 @@
-let index = 0;
+import { fadeOut, fadeIn } from  "./fade.js";
+import { animals } from "./cameraCoordinate.js";
+
 let aframe = document.getElementById("aframe");
+//fade out
 aframe.addEventListener("transitionend", fadeIn);
 let camera = document.getElementById("head");
 let animalIndexName = document.getElementById("navigationIndex");
-aframe.classList.add("aframeTransition");
+let index = 0;
 
-let animals =[
-  {"name": "Lion",
-   "cameraX": 0,
-   "cameraY": 5,
-   "cameraZ": 0
-  },
-  {"name": "Camera 2",
-   "cameraX": -90,
-   "cameraY": 0,
-   "cameraZ": 0,
-  },
-  {"name": "Camera 3",
-   "cameraX": 80,
-   "cameraY": 0,
-   "cameraZ": 0,
-  },
-]
-
-//shift to next camera
+let leftNavigationButton = document.getElementById("leftNavigationButton");
+let rightNavigationButton = document.getElementById("rightNavigationButton");
+leftNavigationButton.addEventListener("click", navigateLeft);
+rightNavigationButton.addEventListener("click", navigateRight);
 
   function updateIndex () {
     console.log(animals[index].cameraX + ", " + animals[index].cameraY + ", " + animals[index].cameraZ);
     animalIndexName.innerHTML = animals[index].name;
     camera.object3D.position.set(animals[index].cameraX, animals[index].cameraY, animals[index].cameraZ);
   }
-
-  function fadeOut() {
-    aframe.style.transition="all 1.5s ease-in-out";
-    aframe.style.filter = "brightness(0%)";
-  }
-  
-  function fadeIn() {
-    aframe.style.transition="all 1.5s ease-in-out";
-    aframe.style.filter = "brightness(100%)";
-  }
-  
+  //moves camera to the next coordinate listed. After last coordinate, camera loops to the first coordinate making it an infinite carousel loop.
   function navigateRight() {
     fadeOut()
     if (index < animals.length-1) {
@@ -51,6 +29,7 @@ let animals =[
       updateIndex();
     }
   }
+  ////moves camera to the previous coordinate listed. After first coordinate, camera loops to the last coordinate making it an infinite carousel loop.
   function navigateLeft() {
     fadeOut();
     if (index > 0) {
