+++ vis/scripts/fade.js
@@ -10,3 +10,14 @@
   }
  */
 
+  function fadeOut() {
+    aframe.style.transition="all 1.5s ease-in-out";
+    aframe.style.filter = "brightness(0%)";
+  }
+
+  function fadeIn() {
+    aframe.style.transition="all 1.5s ease-in-out";
+    aframe.style.filter = "brightness(100%)";
+  }
+
+  export { fadeOut, fadeIn };
