--- vis/scripts/navigateLeft.js
@@ -1,2 +1,0 @@
-//shift to prev camera
-
